requirements.txt
Django>=4.0
djangorestframework>=3.14
ultralytics>=8.0
tensorflow>=2.12
numpy>=1.23
opencv-python>=4.7
Pillow>=9.0
